2020 Day 3
Cleanup Answers for Day 3


Docstrings

+++ get_input.py
@@ -5,11 +5,11 @@ from urllib.request import urlopen
 
 repo_path = Path(__file__).parent
 
-year = 2015
-day = 4
+year = 2020
+day = 3
 url = "http://adventofcode.com/{}/day/{}/input".format(year, day)
 # Cookie for your login info. Each user gets a special series if input data.
-session_id = "53616c7465645f5f2ee10a06bd5c341d4ed52cb913321a608b4de7223122f2f743633271cbbf2039749f30cbfa99f2f7"
+session_id = "53616c7465645f5fb33ae07a0f3e46d59dff717f575af35290b79ec3443771cae3f1b989fe852e399896fdb4e8f954b0"
 
 url_req = urllib.request.Request(url)
 # Add the session id to the request
@@ -29,4 +29,6 @@ os.makedirs(input_path, exist_ok=True)
 # Write out the input data.
 input_file.write_text(response.read().decode("utf-8"))
 # Create the default answer python file.
-answer_file.write_text(answer_temp.read_text())
+if not answer_file.exists():
+    answer_file.write_text(answer_temp.read_text())
+
